requirements.txt
# ffmpeg-python
python-pptx
azure-cognitiveservices-speech